Initialize CVS branch PERL
The per-file CVS branch roots do not exactly match a reconstructed source-branch commit.

+++ gnu/usr.bin/perl/lib/blib.t
@@ -9,27 +9,9 @@ use strict;
 use File::Spec;
 my($blib, $blib_arch, $blib_lib, @blib_dirs);
 
-my $Is_VMS   = $^O eq 'VMS';
-my $Is_VMS_mode = 0;
-
-if ($Is_VMS) {
-    require VMS::Filespec if $Is_VMS;
-    my $vms_unix_rpt;
-
-    $Is_VMS_mode = 1;
-    if (eval 'require VMS::Feature') {
-        $vms_unix_rpt = VMS::Feature::current("filename_unix_report");
-    } else {
-        my $unix_rpt = $ENV{'DECC$FILENAME_UNIX_REPORT'} || '';
-        $vms_unix_rpt = $unix_rpt =~ /^[ET1]/i; 
-    }
-    $Is_VMS_mode = 0 if ($vms_unix_rpt);
-}
-
-
 sub _cleanup {
     rmdir foreach reverse (@_);
-    unlink "stderr";
+    unlink "stderr" unless $^O eq 'MacOS';
 }
 
 sub _mkdirs {
@@ -41,14 +23,25 @@ sub _mkdirs {
     
 
 BEGIN {
-    $blib = "blib";
-    $blib_arch = "blib/arch";
-    $blib_lib = "blib/lib";
-    @blib_dirs = ($blib, $blib_arch, $blib_lib);
+    if ($^O eq 'MacOS')
+    {
+	$MacPerl::Architecture = $MacPerl::Architecture; # shhhhh
+	$blib = ":blib:";
+	$blib_lib = ":blib:lib:";
+	$blib_arch = ":blib:lib:$MacPerl::Architecture:";
+	@blib_dirs = ($blib, $blib_lib, $blib_arch); # order
+    }
+    else
+    {
+	$blib = "blib";
+	$blib_arch = "blib/arch";
+	$blib_lib = "blib/lib";
+	@blib_dirs = ($blib, $blib_arch, $blib_lib);
+    }
     _cleanup( @blib_dirs );
 }
 
-use Test::More;
+use Test::More tests => 7;
 
 eval 'use blib;';
 ok( $@ =~ /Cannot find blib/, 'Fails if blib directory not found' );
@@ -65,13 +58,13 @@ _mkdirs( @blib_dirs );
 is( @INC, 3, '@INC now has 3 elements' );
 is( $INC[2],    '../lib',       'blib added to the front of @INC' );
 
-if ($Is_VMS_mode) {
+if ($^O eq 'VMS') {
     # Unix syntax is accepted going in but it's not what comes out
     # So we don't use catdir above
     $blib_arch = 'blib.arch]';
     $blib_lib = 'blib.lib]';
 }
-else
+elsif ($^O ne 'MacOS')
 {
     $blib_arch = File::Spec->catdir("blib","arch");
     $blib_lib  = File::Spec->catdir("blib","lib");
@@ -80,7 +73,5 @@ else
 
 ok( grep(m|\Q$blib_lib\E$|, @INC[0,1])  == 1,     "  $blib_lib in \@INC");
 ok( grep(m|\Q$blib_arch\E$|, @INC[0,1]) == 1,     "  $blib_arch in \@INC");
-
-done_testing();
 
 END { _cleanup( @blib_dirs ); }
